tested endpoints except admin
